Start 2017.15.2
I _think_ it works as it is, however, it is ridiculously slow. I don't
know why, seeing as the 15.1 version was just fine, but I guess it must
have something to do with the queues, or maybe just because generating
these more controlled values takes a while? I'll have to figure it out
before I can actually "solve" this challenge.