Complete 2020.10.2
This one I was unable to do alone. I got some hints that it could be
solved as a tribonacci sequence (which is like fibonacci but... three),
and that "Dynamic Programming" could help, and this is where I ended up.
I guess that's good though cus it means I learned something.