Improve 2020.11
They each take a little less than half the time now, with the 2D array
swapped out for a single dict. I'm sure there are ways I could get
further improvements, but it may not be worth the extra time at this
point. On the plus side it also looks a bit cleaner. :)