Start 2020.13.2
I'm not really even sure if it works or not. It seems to continue forever,
which sorta makes sense cus departures() does but like one would hope the
code would eventually reach the return in the for loop. It could either be
that something in the logic is wrong, or that this approach is simply too
slow when you're looking for a value as high as we are. To be determined.