Complete 2020.13.2
I owe a lot of thanks to tmdesigned on Reddit for this, as my
implementation is little more than a translation of theirs into Python.
I'm not even confident I learned a lot from it, except maybe a bit about
reading mathematical notation. Probably one of my weaker points.

Here's that Reddit post, by the way:
https://www.reddit.com/r/adventofcode/comments/kc4njx/2020_day_13_solutions/gfswxva/